adds useSavePrivateKey hook and share between extension/mobile
