+++ packages/app-extension/src/components/common/Account/PrivateKeyInput.tsx
@@ -1,12 +1,8 @@
 import { useEffect, useState } from "react";
 import type { Blockchain, ServerPublicKey } from "@coral-xyz/common";
-import {
-  formatWalletAddress,
-  UI_RPC_METHOD_FIND_SERVER_PUBLIC_KEY_CONFLICTS,
-  validatePrivateKey,
-} from "@coral-xyz/common";
+import { formatWalletAddress } from "@coral-xyz/common";
 import { PrimaryButton, TextInput } from "@coral-xyz/react-common";
-import { useAllWallets, useBackgroundClient } from "@coral-xyz/recoil";
+import { useSavePrivateKey } from "@coral-xyz/recoil";
 import { Box } from "@mui/material";
 
 import { Header, SubtextParagraph } from "../../common";
@@ -34,13 +30,14 @@ export const PrivateKeyInput = ({
   displayNameInput?: boolean;
   onboarding?: boolean;
 }) => {
-  const background = useBackgroundClient();
-  // eslint-disable-next-line
-  const wallets = onboarding ? [] : useAllWallets();
   const [name, setName] = useState("");
   const [privateKey, setPrivateKey] = useState("");
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState<string | null>(null);
+
+  const { handleSavePrivateKey } = useSavePrivateKey({
+    onboarding,
+  });
 
   useEffect(() => {
     // Clear error on form input changes
@@ -49,69 +46,14 @@ export const PrivateKeyInput = ({
 
   const onSave = async (e: React.FormEvent) => {
     e.preventDefault();
-    setLoading(true);
-
-    // Do some validation of the private key
-    let _privateKey: string, _publicKey: string, _blockchain: Blockchain;
-    try {
-      ({
-        privateKey: _privateKey,
-        publicKey: _publicKey,
-        blockchain: _blockchain,
-      } = validatePrivateKey(privateKey, blockchain));
-    } catch (e) {
-      setLoading(false);
-      setError((e as Error).message);
-      return;
-    }
-
-    if (wallets.find((w) => w.publicKey === _publicKey)) {
-      setError("This wallet is already active and available in your account.");
-      return;
-    }
-
-    // Check if the public key we have is the public key we wanted (if we were
-    // looking for a specific public key)
-    if (serverPublicKeys && serverPublicKeys.length > 0) {
-      setLoading(false);
-      const found = !!serverPublicKeys.find(
-        (s: { publicKey: string; blockchain: Blockchain }) =>
-          s.publicKey === _publicKey && s.blockchain === _blockchain
-      );
-      if (!found) {
-        if (serverPublicKeys.length === 1) {
-          setError(
-            `Incorrect private key for ${formatWalletAddress(
-              serverPublicKeys[0].publicKey
-            )}. The public key was ${formatWalletAddress(_publicKey)}.`
-          );
-        } else {
-          setError(
-            `Public key ${formatWalletAddress(
-              _publicKey
-            )} not found on your Backpack account.`
-          );
-        }
-      }
-    } else {
-      // If we aren't searching for a public key we are adding it to the account,
-      // check for conflicts.
-      const response = await background.request({
-        method: UI_RPC_METHOD_FIND_SERVER_PUBLIC_KEY_CONFLICTS,
-        params: [[{ blockchain: _blockchain, publicKey: _publicKey }]],
-      });
-
-      if (response.length > 0) {
-        setError("Wallet address is used by another Backpack account");
-        return;
-      }
-    }
-
-    onNext({
-      blockchain: _blockchain,
-      publicKey: _publicKey,
-      privateKey: _privateKey,
+    await handleSavePrivateKey({
       name,
+      privateKey,
+      blockchain,
+      serverPublicKeys,
+      onNext,
+      setLoading,
+      setError,
     });
   };
